refactor: 캔버스 요소 동기화 로직 분리
- syncPostitsToState, syncPlaceCardsToState, syncLinesToState, syncTextBoxesToState, syncZIndexOrderToState를
- 별도의 파일로 분리하고 기존 의존성을 유지하고자 팩토리 패턴으로 진행
